App_Backup/tv-binance-bot/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Template engine
jinja2==3.1.2

# Database ORM
sqlalchemy==2.0.23

# Binance API client
python-binance==1.0.19

# HTTP client for requests
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Date/time utilities
python-dateutil==2.8.2

# OAuth authentication
authlib==1.3.0
itsdangerous==2.1.2

# Excel export
openpyxl==3.1.2

